requirements.txt
# Database connectivity
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0

# Data processing
pandas>=1.5.0

# Visualization
matplotlib>=3.5.0

# Configuration
python-dotenv>=0.19.0
